Fix appending of FASTQ files by sample across lanes
The original, Python-based implementation of the BaseSpace CLI retrieved directories of FASTQ files named with the format `SampleName_SampleId`.  However, the newer, executable version of the CLI retrieves directories of FASTQ files named with the format `SampleName_LaneId_ds.DownloadSessionId`, where LaneId is the glob "L00[1-4]" and DownloadSessionId is a 32-character hex string that is unique to a specific download session (i.e., files from a specific sample/lane combination).  This edit fixes a bug in the code to append FASTQ files across lanes (which had assumed that the only "L00?" glob occurred in the base FASTQ filename) and also updates the comments to reflect the newer directory structure.