Update run ID regex
Starting with the NextSeq 2000, the IDs of sequencing runs generated by the BU Microarray and Sequencing Resource have no longer included the initial "A" or "B" (which indicated the flowcell position on dual-flowcell machines like the HiSeq 2000) before the 9-character flowcell ID.  For this reason, this commit updates the regex to make the A/B position indicator optional.

While this change is being made, the initial part of the regex is also being changed to explicitly reflect the yymmdd format of the sequencing run date.